Add Monster Attack Logic
Add Monster Attack Logic

+++ Assets/2.Scripts/Scriptable/MonsterDataScriptable.cs
@@ -11,12 +11,12 @@ namespace Game.Data
         [SerializeField] private int maxHp;
         [SerializeField] private float speed;
         [SerializeField] private float range;
-        [SerializeField] private float attackPower;
+        [SerializeField] private int attackPower;
 
         public int MaxHP => maxHp;
         public float Speed => speed;
         public float Range => range;
-        public float AttackPower => attackPower;
+        public int AttackPower => attackPower;
     }
 }
